% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hindex.R
\name{compute_hindex}
\alias{compute_hindex}
\title{Compute 3D Hilbert curve indices}
\usage{
compute_hindex(x, y, z, max_coord_bits)
}
\arguments{
\item{x, y, z}{Integer vectors of coordinates}

\item{max_coord_bits}{Number of bits per dimension (e.g., 10 for a 1024x1024x1024 cube)}
}
\value{
A numeric vector of Hilbert indices.
}
\description{
Compute 3D Hilbert curve indices
}
