% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_hindex_cpp}
\alias{compute_hindex_cpp}
\title{Compute 3D Hilbert curve indices}
\usage{
compute_hindex_cpp(x, y, z, nbits, as_character = FALSE)
}
\arguments{
\item{x, y, z}{Integer vectors of coordinates}

\item{nbits}{Number of bits per dimension (1-21)}

\item{as_character}{Return indices as character strings to preserve precision}
}
\value{
Numeric or character vector of Hilbert indices
}
\description{
Compute 3D Hilbert curve indices
}
\examples{
compute_hindex_cpp(IntegerVector::create(0,1,2),
                   IntegerVector::create(0,1,2),
                   IntegerVector::create(0,1,2), 4)
}
