% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_hindex_cpp_inverse}
\alias{compute_hindex_cpp_inverse}
\title{Decode Hilbert curve indices back to coordinates}
\usage{
compute_hindex_cpp_inverse(index, nbits)
}
\arguments{
\item{index}{Character vector of Hilbert indices}

\item{nbits}{Number of bits per dimension (1-21)}
}
\value{
A data.frame with columns x, y and z
}
\description{
Decode Hilbert curve indices back to coordinates
}
