% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hindex.R
\name{compute_hindex_inverse}
\alias{compute_hindex_inverse}
\title{Decode Hilbert curve indices back to coordinates}
\usage{
compute_hindex_inverse(h_indices, max_coord_bits)
}
\arguments{
\item{h_indices}{A numeric vector of Hilbert indices}

\item{max_coord_bits}{Number of bits per dimension}
}
\value{
A data.frame with columns x, y and z
}
\description{
Decode Hilbert curve indices back to coordinates
}
