% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{dim,ParquetNeuroVec-method}
\alias{dim,ParquetNeuroVec-method}
\title{Get Dimensions of ParquetNeuroVec}
\usage{
\S4method{dim}{ParquetNeuroVec}(x)
}
\arguments{
\item{x}{ParquetNeuroVec object}
}
\value{
integer vector. Dimensions of the 4D array
}
\description{
Get Dimensions of ParquetNeuroVec
}
