% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hilbert3D_single}
\alias{hilbert3D_single}
\title{Convenience wrapper for a single 3D point}
\usage{
hilbert3D_single(x, y, z, nbits, as_character = FALSE)
}
\arguments{
\item{x, y, z}{Integer coordinates}

\item{nbits}{Number of bits per dimension (1-21)}

\item{as_character}{Return index as character string}
}
\value{
A length-one vector containing the Hilbert index
}
\description{
Convenience wrapper for a single 3D point
}
