% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{length,ParquetNeuroVec-method}
\alias{length,ParquetNeuroVec-method}
\title{Get Length of ParquetNeuroVec (Number of Time Points)}
\usage{
\S4method{length}{ParquetNeuroVec}(x)
}
\arguments{
\item{x}{ParquetNeuroVec object}
}
\value{
integer. Number of time points
}
\description{
Get Length of ParquetNeuroVec (Number of Time Points)
}
