% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{series,ParquetNeuroVec,integer-method}
\alias{series,ParquetNeuroVec,integer-method}
\title{Extract Time Series from ParquetNeuroVec}
\usage{
\S4method{series}{ParquetNeuroVec,integer}(x, i, j, k, drop = TRUE)
}
\arguments{
\item{x}{ParquetNeuroVec object}

\item{i}{integer. x-coordinate (1-based)}

\item{j}{integer. y-coordinate (1-based)}

\item{k}{integer. z-coordinate (1-based)}

\item{drop}{logical. Whether to drop dimensions of length 1}
}
\value{
numeric vector of time series values
}
\description{
Extract Time Series from ParquetNeuroVec
}
