% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{series,ParquetNeuroVec,matrix-method}
\alias{series,ParquetNeuroVec,matrix-method}
\title{Extract Time Series Using Matrix Coordinates}
\usage{
\S4method{series}{ParquetNeuroVec,matrix}(x, i)
}
\arguments{
\item{x}{ParquetNeuroVec object}

\item{i}{matrix. Matrix with 3 columns (x, y, z coordinates, 1-based)}
}
\value{
matrix. Each column is a voxel's time series
}
\description{
Extract Time Series Using Matrix Coordinates
}
