% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{series,ParquetNeuroVec,numeric-method}
\alias{series,ParquetNeuroVec,numeric-method}
\title{Extract Time Series from ParquetNeuroVec (Numeric Coordinates)}
\usage{
\S4method{series}{ParquetNeuroVec,numeric}(x, i, j, k, drop = TRUE)
}
\arguments{
\item{x}{ParquetNeuroVec object}

\item{i}{numeric. x-coordinate}

\item{j}{numeric. y-coordinate}

\item{k}{numeric. z-coordinate}

\item{drop}{logical. Whether to drop dimensions}
}
\value{
numeric vector of time series values
}
\description{
Extract Time Series from ParquetNeuroVec (Numeric Coordinates)
}
