% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetNeuroVec.R
\name{series_roi_bbox,ParquetNeuroVec,integer,integer,integer-method}
\alias{series_roi_bbox,ParquetNeuroVec,integer,integer,integer-method}
\title{Extract ROI Time Series from ParquetNeuroVec based on a bounding box}
\usage{
\S4method{series_roi_bbox}{ParquetNeuroVec,integer,integer,integer}(x, i, j, k)
}
\arguments{
\item{x}{ParquetNeuroVec object}

\item{i}{integer. x-coordinate range or single coordinate}

\item{j}{integer. y-coordinate range or single coordinate}

\item{k}{integer. z-coordinate range or single coordinate}
}
\value{
numeric vector. Average time series across the ROI
}
\description{
Extract ROI Time Series from ParquetNeuroVec based on a bounding box
}
